+++ NEWS.md
@@ -1,3 +1,7 @@
+# sharpshootR 1.10 (2022-06-19)
+ * adding SoilTaxonomy package to suggests, to allow for better encoding of taxa levels by `SoilTaxonomyDendrogram`
+ * `SoilTaxonomyDendrogram()` gains argument `KST.order` to adjust encoding / ordering criteria
+
 # sharpshootR 1.9.1 (2022-02-16)
  * bugfix in `plotWB()` related to figure y-scale range
  * `plotWB_lines()` gets `legend.cex` and aesthetic adjustments to legend position